Huge commit! BEWARE! Migrate to ASP.NET Core. Migrate Settings to new modular system.

+++ Shoko.Server/API/v2/Models/core/Credentials.cs
@@ -4,10 +4,10 @@
     {
         public string login { get; set; }
         public string password { get; set; }
-        public int port { get; set; }
+        public ushort port { get; set; }
         public string token { get; set; }
         public string refresh_token { get; set; }
         public string apikey { get; set; }
-        public int apiport { get; set; }
+        public ushort apiport { get; set; }
     }
 }
